Moved component_pools class into its own header
